Move playground app to website folder
