--- playground/firebase.json
@@ -1,18 +1,0 @@
-{
-    "hosting": {
-        "target": "playground",
-        "public": "public",
-        "ignore": [
-            ".firebaserc",
-            "firebase.json",
-            "**/.*",
-            "**/node_modules/**"
-        ],
-        "rewrites": [
-            {
-                "source": "/**",
-                "destination": "/index.html"
-            }
-        ]
-    }
-}
