+++ website/playground/actions/playground.js
@@ -5,9 +5,6 @@ import lzString from "lz-string";
 const compressStr = str => lzString.compressToBase64(str);
 const decompressStr = str => lzString.decompressFromBase64(str);
 
-const htmlSymbol = "siimple:playground:html";
-const configSymbol = "siimple:playground:config";
-
 // Load playground content from source
 export const loadPlayground = content => {
     return new Promise((resolve, reject) => {
@@ -46,7 +43,8 @@ export const sharePlayground = content => {
         // }
         query.set("html", compressStr(content.html));
         query.set("config", compressStr(content.config));
-        return `${window.location.origin}#${query.toString()}`;
+        // return `${window.location.origin}#${query.toString()}`;
+        return `${window.location.href}#${query.toString()}`;
     });
 };
 
@@ -62,32 +60,4 @@ export const exportPlayground = content => {
     }));
     const fileName = `playground-${kofi.timestamp("YYYY-MM-DD")}.json`;
     return kofi.downloadFile(fileName, fileContent);
-};
-
-// Update preview hanlder
-export const inject = (preview, html, css) => {
-    const data = {
-        source: "siimple-playground",
-        html: html,
-        css: css,
-    };
-    return preview.contentWindow.postMessage(data, "*");
-};
-
-// Compile siimple
-export const compile = (worker, config) => {
-    return new Promise(resolve => {
-        const id = kofi.tempid();
-        const onMessage = event => {
-            if (event.data?.id === id) {
-                worker.removeEventListener("message", onMessage);
-                return resolve(event.data);
-            }
-        };
-        worker.addEventListener("message", onMessage);
-        worker.postMessage({
-            id: id,
-            config: config,
-        })
-    });
 };
