+++ website/playground/hooks/useEditor.js
@@ -2,9 +2,7 @@ import React from "react";
 import {CodeCake} from "codecake/codecake.js";
 import {highlight} from "codecake/highlight.js";
 import {lineNumbers} from "codecake/linenumbers.js";
-import {defaultHtml, defaultConfig} from "./defaults.js";
 
-// Editor hook
 export const useEditor = (parent, lang) => {
     const editor = React.useRef(null);
     React.useEffect(() => {
@@ -15,19 +13,4 @@ export const useEditor = (parent, lang) => {
         }
     }, []);
     return editor;
-};
-
-// Playground hook
-export const usePlayground = () => {
-    const playground = React.useRef(null);
-    if (!playground.current) {
-        playground.current = {
-            version: "latest",      // siimple version 
-            html: defaultHtml + "\n",
-            config: defaultConfig + "\n",
-            hasHtmlChanges: false,
-            hasConfigChanges: false,
-        };
-    }
-    return playground;
 };
