+++ website/playground/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@siimple/playground",
+    "name": "siimple-playground",
     "version": "1.0.0",
     "private": true,
     "description": "Online editor for siimple",
@@ -9,11 +9,10 @@
     "repository": "https://github.com/jmjuanes/siimple",
     "bugs": "https://github.com/jmjuanes/siimple/issues",
     "scripts": {
-        "clean": "rm -rf public && rm -rf .cache",
         "build": "webpack build --config ./webpack.config.js",
-        "serve": "stattic --folder ./public --port 5000"
+        "clean": "rm -rf public"
     },
-    "devDependencies": {
+    "dependencies": {
         "@babel/core": "^7.17.8",
         "@babel/plugin-transform-react-jsx": "^7.17.3",
         "@babel/plugin-transform-runtime": "^7.17.0",
@@ -27,7 +26,6 @@
         "lz-string": "^1.4.4",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
-        "stattic": "^0.4.0",
         "webpack": "^5.69.0",
         "webpack-cli": "^4.9.2"
     }
